--- src/jmh/java/BaseBenchmark.java
@@ -1,14 +1,0 @@
-package edu.java.fintechcourse2024.hw15;
-
-import org.openjdk.jmh.annotations.*;
-
-import java.util.concurrent.TimeUnit;
-
-@BenchmarkMode(Mode.Throughput)
-@OutputTimeUnit(TimeUnit.SECONDS)
-@State(Scope.Thread)
-@Fork(1)
-@Warmup(iterations = 1)
-@Measurement(iterations = 3)
-public abstract class BaseBenchmark {
-}
