+++ src/jmh/java/kafka/test/SimpleConfigurationBenchmark.java
@@ -1,6 +1,6 @@
-package edu.java.fintechcourse2024.hw15.kafka.test;
+package kafka.test;
 
-import edu.java.fintechcourse2024.hw15.kafka.KafkaProducerConsumerBenchmarkBase;
+import kafka.KafkaProducerConsumerBenchmarkBase;
 
 public class SimpleConfigurationBenchmark extends KafkaProducerConsumerBenchmarkBase {
     public SimpleConfigurationBenchmark() {
