+++ src/jmh/java/kafka/test/StressTestBenchmark.java
@@ -1,6 +1,6 @@
-package edu.java.fintechcourse2024.hw15.kafka.test;
+package kafka.test;
 
-import edu.java.fintechcourse2024.hw15.kafka.KafkaProducerConsumerBenchmarkBase;
+import kafka.KafkaProducerConsumerBenchmarkBase;
 
 public class StressTestBenchmark extends KafkaProducerConsumerBenchmarkBase {
     public StressTestBenchmark() {
