+++ src/jmh/java/rabbit/RabbitMQConsumer.java
@@ -1,10 +1,9 @@
-package edu.java.fintechcourse2024.hw15.rabbit;
+package rabbit;
 
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
 import com.rabbitmq.client.DeliverCallback;
-import lombok.Getter;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
