+++ src/jmh/java/rabbit/test/StressTestBenchmark.java
@@ -1,6 +1,6 @@
-package edu.java.fintechcourse2024.hw15.rabbit.test;
+package rabbit.test;
 
-import edu.java.fintechcourse2024.hw15.rabbit.RabbitMQBenchmarkBase;
+import rabbit.RabbitMQBenchmarkBase;
 
 public class StressTestBenchmark extends RabbitMQBenchmarkBase {
     public StressTestBenchmark() {
